pyproject.toml
[project]
name = "nonebot-plugin-osugreek"
version = "1.0.9"
description = "在图片中央贴上神秘4k希腊字母并添加色散效果的 NoneBot2 插件"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"

dependencies = [
    "nonebot2>=2.3.0",
    "nonebot-adapter-onebot>=2.0.0",
    "pillow>=9.0.0",
    "aiohttp>=3.8.0",
    "nonebot-plugin-localstore>=0.3.0",
]

[project.urls]
Homepage = "https://github.com/YakumoZn/nonebot-plugin-osugreek"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"nonebot_plugin_osugreek" = ["images/*.png"]

[project.entry-points."nonebot.plugin"]
osugreek = "nonebot_plugin_osugreek"